+++ src/vocabulous/__init__.py
@@ -16,6 +16,6 @@ Example:
 
 from .vocabulous import Vocabulous
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __author__ = "Omar Kamali"
 __all__ = ["Vocabulous"]